import.php: Only import objects.cache when needed
In Bad Situations, the import script will be run without nagios having
been restarted. In those cases, the import script would have immediately
truncated the database, causing ninja to be useless until the script has
finished.

With this change, objects.cache will only be read if it has changed
since the last run, roughly halving the execution time in those cases.
Also, we now avoid truncating the database for as long as we can, so if
the import takes a long time to run, the time when ninja is useless
should be significantly decreased.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>