Makefile: Avoid linking directly to oracle libraries
The oracle client library rpm's are flawed, since they do not
provide the necessary "provides" lines for the libraries they
include.

ocilib, which merlin uses, is in our environment compiled to
load the oracle libraries at runtime rather than at linktime,
so we really don't need to link anything to oracle directly.

This solves the problem quite neatly and also makes it
possible for us to ship an oracle-enabled merlin daemon by
default even to people who do not want to use or install the
oracle client libraries.

Signed-off-by: Andreas Ericsson <ae@op5.se>