specfile: Make merlin require ocilib >= 3.8
This is what we test against anyways, so it makes sense to make
that the earliest possible version to use. It also means it
will automatically get pulled in when upgrading a Merlin that's
linked to it (although that should've happened anyway since the
library provides are automatic with RPM).

Signed-off-by: Andreas Ericsson <ae@op5.se>