import.php: Work around erroneously buggy ocilib string quoter
For some (unknown) reason, the ocilib string quoter appears to be
buggy and returns 'false' instead of a quoted string, no matter
what is passed to it. This patch works around it in the quickest
and dumbest way possible. Since it's highly unlikely that we'll
get attack-type strings to quote, this function should work
well enough.

Signed-off-by: Andreas Ericsson <ae@op5.se>