sql: Make sql_is_connected() optionally try reconnecting
Everywhere where it's used is a reconnect point if we're not already
connected, so we'd better try establishing the connection from there.
We do want to re-use it from inside sql_init() though, so the reconnect
thing has to be optional if we're to avoid an infinite loop that will
quickly eat all available stackspace.

Signed-off-by: Andreas Ericsson <ae@op5.se>