daemon: Send CTRL_STALL a bit more frequently
With an allowed interval of 20 seconds from the module's point of
view and a configured interval of 15 seconds inside a loop that
has multiple other timeout spots makes it possible that the module
starts sending events before the daemon is ready to accept them.

This gives a very visible impact when running testcases and the
mtest program starts shipping events before the daemon is done
importing, which causes a lot of tests to fail.

Signed-off-by: Andreas Ericsson <ae@op5.se>