misc: Make getcwd(3) usage Solaris compatible
The previous code would bomb out horribly on Solaris when the
cfg_dir directive was encountered, causing a sigsegv when trying
to use the returned value of the getcwd(3) call, but since we
called it with a zero path-length the solaris libc version issued
a zero-length malloc() to store the data and promptly returned
NULL.

This patch amends it by putting the buffers on the stack and
passing them as the first argument to getcwd(3), with their
actual size as the second argument.

This means we now have a limitation of PATH_MAX in how deep a
directory structure we can handle, but since PATH_MAX is generally
4096 bytes we should be fairly safe for all sane usecases.

As a nice bonus, we no longer leak the few bytes of memory we
earlier did when invoking the get_cfg_files() helper when a
cfg_dir directive is specified in the nagios.cfg config file.

Signed-off-by: Andreas Ericsson <ae@op5.se>