daemon: Always destroy the sql connect remnants when we fail to connect
With the addition of blocked logspamming we had managed to also cause
the sql_init() function to not clean up after itself, making the
rest of the daemon believe it had a valid database connection when
this wasn't the case. This patch takes care of it by moving the
sql_close() statement outside of the if(log_attempt){} block.

Signed-off-by: Andreas Ericsson <ae@op5.se>