mon: Make merlin_db module return database connection object
It used to return a cursor, but that caused some problems since
the caller had no option of getting the raw connection right
from the get-go, which in turn prevented us from using the database
to its fullest.

This approach is better, since callers can still get the cursor
directly by just appending .cursor() to the merlin_db.connect()
call.

Signed-off-by: Andreas Ericsson <ae@op5.se>