Break 'IN()' clauses into elements of 1000 when deleting objects
Oracle has a hard limit of 1000 elements in IN clauses. It appears
there's no upper limit for the number of IN clauses in the query
though, so we can safely split them into AND-joined IN() clauses of
1000 elements each.

In order to exercise the code properly, we use the same code on MySQL,
with no noticeable slowdown.

Signed-off-by: Andreas Ericsson <ae@op5.se>