install: Don't run add index for permanent tables
Installing new indexes for massive tables is hysterically slow, and some
merlin users have accumulated massive tables. To prevent the upgrade
from unexpectedly locking all tables causing merlin not to function, we
avoid installing indexes for the permanent tables during upgrades.

If the user is currently lacking indexes for permanent tables, we shout
during the install. It's still easy to miss, but it's better than
nothing.

Also add a mon command for the user to run during a downtime window to
add the missing indexes.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>