object_importer: Fix object deletion
This was actually three separate bugs, but tightly coupled.

PHP doesn't accept + as a string concatenator but instead turns the
entire string into numeral 0, which is patently absurd. This bug hid
the two others, since '0' is apparently valid SQL and doesn't generate
an error.

I had also forgotten the ',' part of the join() call, causing all
indices to be concatenated as one long string. Nice.

On top of that, I had fubar'ed the query completely by always making
the where-clause start with "id NOT IN id NOT IN".

I'm really, really proud of myself now. Awesome.

Signed-off-by: Andreas Ericsson <ae@op5.se>