daemon: Fix typo in fcntl(2) call
We used to ask for F_SETFD when we really meant F_SETFL, since we
want to set file descriptor status flags and not file descriptor
flags. I don't know why this didn't throw an error, but it's possible
the O_NONBLOCK flag collides with some value valid for F_SETFD actions
too.

This should prevent Merlin from hanging for a long time when one or
more configured nodes which we should try to connect to are behind
firewalls or just plain misconfigured.

Signed-off-by: Andreas Ericsson <ae@op5.se>