net: Avoid connecting when already connected
Currently we're spamming already connected nodes with connection
attempts, which is clearly stupid. We're far better off avoiding
the extra overhead by skipping the connection attempt at all.

We also modify the error handling when we get EISCONN when trying
to connect so we don't even bother with the whole getpeername(2)
business.

Signed-off-by: Andreas Ericsson <ae@op5.se>