module: Revamp how we get the selection id for host and service commands
Previously we used to require a semi-colon in the command arguments,
but with this patch we don't. This makes "blank" host commands work
as they should again (such as "DISABLE_HOST_PASV_CHECKS" or whatever
the command is). Previously they were blocked from pollers since they
didn't have a semi-colon after the hostname in the command arguments.

Signed-off-by: Andreas Ericsson <ae@op5.se>