module: Special-case hostgroup commands
Since hostgroups are used as selections and we know which poller or
group of pollers handles a particular hostgroup, we can and should
direct commands regarding such hostgroups to that particular (group
of) poller(s). This patch makes it so and also prevents commands for
hostgroups sharing name with a host from ending up on the wrong
poller.

Signed-off-by: Andreas Ericsson <ae@op5.se>