module: Block notifications on nodes that aren't handling the object
When a notification-triggering command is submitted to a node not
normally responsible for handling a particular host or service check,
a notification is sent once per peer in each layer concerned. This
patch blocks notifications from any nodes but the one that is currently
supposed to handle the checking of some particular node, thus preventing
multiple notifications from being sent.

Signed-off-by: Andreas Ericsson <ae@op5.se>