hooks/net: Don't special case comments
Always send all comments, and not just acknowledgements, from nebmodule
to daemon, and add them to db in the daemon only when they arrive from
the local socket.

Nagios already only adds one comment object per logical comment, but
when we added comments both from the net and from nagios, we got
duplicates. This was handled by trying to avoid forwarding comments,
however, that made them never arrive on pollers.

This does create some ping-pong action, but very limited, and in an area
that is not bandwidth critical.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>