daemon: Remove comments a bit more cleverly
Previously we used to rely on comment_id when getting rid of
comments, but due to Nagios' way of not synchronizing comment
id's we can't really do that and know we're deleting the right
comment. Instead we rely on entry_time, object name, author name
and comment_data to locate the right comment. This means that
identical comments submitted on the same second will be deleted
together, but that's probably not a bad thing.

Signed-off-by: Andreas Ericsson <ae@op5.se>