module: Stash id's of flapping comments when adding them
Flapping comments are not sent as command events, but only as
comment events. In order to keep track of the comment id locally
so we transmit the proper information in case of check failover,
we need to set the local comment_id properly in the monitored
object. We solve this by creating a comment locally and saving
the id of that comment in the node struct for the sending node.
This lets us set the comment id properly when we receive the
flapping start event shortly after.

We also add a stub for "add_new_comment()" in nagios-stubs.h so
that mtest links properly.

Signed-off-by: Andreas Ericsson <ae@op5.se>