daemon: Avoid sending events magically marked as NONET events
We add the MAGIC_NONET marker to be used in the merlin_header.code
field for non-control events that shouldn't be transferred across
the network. This makes it possible to avoid ping-pong action
between several nodes when handling comments, which need to pass
through the module to be inserted with the local comment id in the
local database to avoid the possibility of deleting the wrong comment.

Signed-off-by: Andreas Ericsson <ae@op5.se>