module: Once again handle acknowledgements by commands
Acknowledgements trigger notifications, but the notification
hook prevents notifications from being sent out from nodes
other than the one supposed to run the check for the host
or service in question, but when we moved acknowledgements
to be handled like other comments and blocked their commands,
we prevented any notification from being sent at all.

This patch remedies that, special-casing acknowledgement
comment events so we mark them with MAGIC_NONET and prevent
undue ping-pong bouncing and reinstating the forwarding of
acknowledgement commands instead.

Signed-off-by: Andreas Ericsson <ae@op5.se>