module: Special-case DOWNTIME type comments
Scheduled downtime is also forwarded as commands (and that's
the way we want it to be), so there's no real need for us to
send the comment events for them beyond the reach of our own
local Merlin module.

Previously one downtime comment was added once per connected
node capable of monitoring the object we scheduled downtime
for. This patch makes it so that only one downtime comment is
ever entered for any particular downtime on any particular
object.

Signed-off-by: Andreas Ericsson <ae@op5.se>