mon node ctrl: Don't require --all when passing other arguments
The earlier check was quite weird in that it required us to
pass --all if we wanted the command to be run on all nodes
when also passing --self or --merlin-cfg=<something>. This
patch makes a lot more sense, setting the desired target
nodes to "all" if no node names were passed, regardless of
running the command on --self or not.

Since it's patently absurd to run the command *only* on
--self using "mon node ctrl", this makes much more sense and
saves on both typing and irritation.

Signed-off-by: Andreas Ericsson <ae@op5.se>