mtest: Revamp the sleeper function
  MTEST_SLOW=p ./mtest <config>

now makes mtest await one pressing the enter key where it used to
expect a signal that it never caught anyway. It also makes the
sleeptime somewhat configurable by a invocation like this:

  MTEST_SLOW=7 ./mtest <config>

which makes it sleep() 7 seconds instead of the default usleep()
interval. Using commit_interval * 2 + 1 seems to be a sensible formula
for the sleeptime.

Signed-off-by: Andreas Ericsson <ae@op5.se>