object_importer: Avoid using autocommit even with MySQL
When using MySQL and InnoDB tables, the database runs a sanity check
on all tables and makes sure to flush them to disk with each commit.
When using autocommit on a large installation, this causes significant
disk I/O overhead and makes the MySQL server consume copious amounts
of CPU. This patch largely alleviates the problem by committing queries
in chunks. More testing is probably needed to make sure we don't try
to commit too large transactions, but this should work splendidly for
now.

Timings done with 46 hosts and 372 services, best out of 5 runs on
MySQL server with InnoDB tables.

With patch: 17.6 seconds.
Without patch: 2.8 seconds.

Signed-off-by: Andreas Ericsson <ae@op5.se>