Add ocimp
ocimp (Object Config IMPorter) is a complete rewrite of import.php
in C. This makes things go quite a bit faster. An import that runs
in > 6 minutes earlier completes in ~2.5 minutes now, with the same
result in the database.

The most timeconsuming part of the example import is still creating
the contact_access cache, which now consumes a full 75% of the
total runtime. To be fair though, it causes more than 20 times the
amount of queries to be run too.

Signed-off-by: Andreas Ericsson <ae@op5.se>