ocimp: Make --no-sql actually do something
When we're dry-running the program to test the algorithm as if we're
using an empty database we don't want to pay the cost of actually
running the queries. This patch avoids that, and also makes the code
more robust to query failures for random reasons when trying to load
instance_id's and id's already present in the database.

Signed-off-by: Andreas Ericsson <ae@op5.se>