Don't claim to be sending control packets before we actually send them
Previously the check and debug logging was done in node_send_event(),
which can divert events to either a socket (if sending has a chance
of being successful) or to the binlog (if it can't). This caused a
lot of redundant messages along the lines of "Sending CTRL_X to Y"
which looked quite ridiculous when the line above claims that Y has
in fact been disconnected for the past 20 minutes.

Since the latest patch that blocks certain CTRL_ type packets from
entering the binlog the log entries were almost always completely
false as well, so that makes two thumbs up for this change.

Signed-off-by: Andreas Ericsson <ae@op5.se>