daemon: log and disconnect a bit heavier in net_is_connected()
It doesn't make sense not to use the node pointer when acting on
the nodes themselves, and we'll want to make sure the current
state is either PENDING or CONNECTED before we return any value.

If the state is PENDING and we get ENOTCONN we only disconnect
the node if enough time has passed. The default we set to wait
is 20 seconds. The timeout is now defined as a macro and re-used
as send and receive timeout for all asynchronous network ops.

Signed-off-by: Andreas Ericsson <ae@op5.se>