Revamp how node action handlers work
Since some of the actions taken by the action handlers trigger
events to be sent that will cause connection attempts or checks
if the node is actually connected or not, we need to make sure
we mark the node with the new state before running any handlers.
This makes the node->state flag far more reliable than it used to
be and we can now relax a few checks in some of the data shipping
functions.

While we're at it, we remove the check to see if the previous state
matches the new state. We know they won't, since that check is
already run in the node_set_state() function, which is the only
caller for the node action handlers.

Signed-off-by: Andreas Ericsson <ae@op5.se>