daemon: Short-circuit the net_is_connected() logic
Since we'll mark the node as disconnected first error we run into,
we might as well take the shortcuts where we can and assume that
a socket descriptor greater than 0 and a node->state marking the
node as connected means it really is connected. This also means
we assume that we don't even need to bother checking the node
for connectedness if node->state isn't STATE_PENDING, which is
as it should be.

The precedence rules of logical OR's makes it impossible to
write as anything but three if-statements while retaining
readability though. Shame that, really.

Signed-off-by: Andreas Ericsson <ae@op5.se>