ocimp: Add missing slist_sort() call to timeperiod slist
Without this it fails to properly parse timeperiods when the
alphabetical sorting is out of whack for objects.cache. That
should happen very rarely, but with different character sets
or with a hand-mangled objects.cache file we'll run into it
sooner or later for sure.

This patch also merges some other near-identical changes I
made to support excludes in timeperiods. It seems great minds
really do think alike :)

Signed-off-by: Andreas Ericsson <ae@op5.se>