Make tv_delta() stop printing negative seconds
A minor irk, but an annoying one. While we're at it, we get rid
of the floating point math and add weeks as the largest possible
unit. It is, since month and years are quite untrustworthy with
regards to how long they are in seconds. And yes, We conveniently
ignore leapseconds.

Signed-off-by: Andreas Ericsson <ae@op5.se>