ocimp: Fix import of contacts
Previously we used to insert {host,service}_notification_period
as strings, but the database wants them as indices into the
timeperiod table. MySQL didn't balk, although it automagically
convert the timeperiod "24x7" to "24" and inserted everything
not beginning with a digit as "0", but the importer breaks down
horribly on oracle and, presumably, other input-sane databases
without this change.

Signed-off-by: Andreas Ericsson <ae@op5.se>