module: Prevent multiple threads from running hooks at the same time
Otherwise a checkresult event could be running through the module at
the same time as an event triggered by an external command. That's
really only a problem if the ipc binlog has entries, but if it does
we need to make sure only one hook at the time is running.

We could make this lock far more efficient by only invoking it if
the binlog has entries and we're trying to send them, but since the
number of crashes we've seen from this error is very (very) low,
it's probably safe to assume that there will be very little lock
contention.

Signed-off-by: Andreas Ericsson <ae@op5.se>