module: Re-assign peer ids when we receive new node info
Since the handle_ctrl_active() returns 0 if the info is new, we can
make sure we never miss re-assigning peer-id's by the rather simple
expedient of always doing it when the peer gets updated. This is
necessary to make peer id assignment work well again since we resorted
to using node_set_state() in the module when a node connects but before
it's sent a CTRL_ACTIVE packet.

This also lets us force peer id dancing by using gdb to alter
variables inside a running module on one node.

Signed-off-by: Andreas Ericsson <ae@op5.se>