module: Logspam avoidance patch
When a node sends (event) data before it's sent any CTRL_ACTIVE packet,
we rightly print a warning. Most of the times, however, there's more
than one packet arriving from the offending node, and we don't want to
log the message more than once. This patch achieves that end by using
the node->info.byte_order field as a magic marker to get rid of a lot
of annoying messages.

Signed-off-by: Andreas Ericsson <ae@op5.se>