ocimp: Handle multiple running instances
ocimp does not lock tables, thus it is vulnerable to races where two
instances work on the same table at once. As contact_access often takes
about 1/2 of the total runtime, the race will probably happen there,
so that the newest instance truncates it and starts to fill it up while
the old process is still filling it up as well. This results in
duplicate entries in contact_access, causing duplicate entries in ninja.

Instead of doing it the old way by simply locking the tables and waiting
for the old process to complete before ruining all it's hard work by
starting over, we kill the old process before starting ourselves. The
exception is if object.cache hasn't changed, in which case we must wait
for the old process to finish it's (possibly) complete import before
doing our partial import.

We're being a bit extra anal about races, as two importers running at
the exact same time tends to result in loads of SQL errors and little
else.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>