mon oconf: Support syncing arbitrary files and directories between nodes
We do this by adding a 'sync' section to each peer or poller, which
can be populated with raw paths and one flag (restart = yes/no), to
indicate if Nagios and Merlin on the receiving end needs to be restarted
after the synchronization. The extra paths are synced one by one *after*
the object config is sent, which users need to be aware of.

Signed-off-by: Andreas Ericsson <ae@op5.se>