ocimp: Always truncate some ancillary tables
This is necessary in order to make sure we never end up in the
situation where a user has deleted all their hostdependencies
(for example) and then re-run ocimp. Previously, we would leave
all the old hostdependencies in the database. With this patch,
we always remove them no matter what, sacrificing partial UI
availability for the sake of correctness.

Signed-off-by: Andreas Ericsson <ae@op5.se>