slist: Force-sort when seek'ing on an unsorted list
This should make the slist api a lot more predictable and useful,
since apps reading data in mostly sequential order will just take
a performance-hit if it tries to do lookups while inserting stuff
in the sorted list in random order. If the caller knows that's what
he'll be doing, we'd be better off handling this with insertion-sort,
but we'll implement that when it's actually needed.

Signed-off-by: Andreas Ericsson <ae@op5.se>