ocimp: Support host and serviceescalations properly
Previously we never even inserted them into the database, which was
kinda bogus. With this patch we do that and also support the quick
way of caching contact access rights, even though we cheat quite a
bit and insert each individual contact assigned to escalations by
contactgroup affiliation to the {host,service}escalation_contact
table in order to make the final query a lot easier.

It's possible this can be speeded up even more by going the extra
mile and doing the multi-join, multi-union query required to handle the
more complex {host,service}escalation_contacgroup junctions relations.

This is enough improvements for now though, so we'll just let this one
roll out and await feedback.

Signed-off-by: Andreas Ericsson <ae@op5.se>