ocimp: Stash preloaded contact id's in a separate slist
Otherwise we take quite a large and annoying performance hit when
someone adds a lot of new contacts, since the list will be unsorted
most of the times we try to look one up after the preloading, so the
slist api will force-sort the list for us.

This patch gets rid of that annoying scenario, causing us to only sort
either contact slist once.

Signed-off-by: Andreas Ericsson <ae@op5.se>