ocimp: Avoid needlessly allocating object-specific contact slists
Since we'll most often be using the giant query to handle contact
access caching nowadays, we should avoid allocating memory for
old-school access as much as possible. This patch does just that.

Once the new code has seen some more testing we should be able to
get rid of the old one completely.

Signed-off-by: Andreas Ericsson <ae@op5.se>