ocimp: Never truncate more than we have to
When objects.cache hasn't been changed, we'll only ever update the
host and service tables, but previously we mercilessly truncated
them all.

This patch hoists the "always insert this object type" condition out
of the inner loop and truncates all tables that will change at the
start of the run, but only if we will actually write to those tables.

This should fix the problem with empty host- and service-group
pages that some early beta-testers have seen.

Signed-off-by: Andreas Ericsson <ae@op5.se>