cfgfile: Grok variable lines ending with '}'
Previously we took such variables to be compound closers, which was
kinda stupid. Now, the only way left to close compounds is to set
the closing curly brace as the first char of the line, which is
almost certainly the only sane way of doing it, unless we want to
parse char-by-char (which we sort of do anyway...).

Signed-off-by: Andreas Ericsson <ae@op5.se>