mon oconf: Remove dead code
We can resurrect it from git later if we want to, so it makes no sense
to keep ancient, poorly commented code that has never been properly
exercised and is almost certainly buggy. The import parts have been
replaced by ocimp as well, so we'll most likely never need them.

Signed-off-by: Andreas Ericsson <ae@op5.se>