mon oconf: Support all slave type objects somewhat properly
"somewhat" because we cheat like mad and assign even service slaves to
the hosts. This works fine for us, since config is split by hosts
anyways, and it means we can reuse the service assignment code for
serviceescalations and servicedependencies as well.

This is still not very well tested and might break if there are cross-
node dependencies (well, it *will* break then, actually), but for now
this should make sure that escalations and dependencies get sent to the
right node at least.

Signed-off-by: Andreas Ericsson <ae@op5.se>