Add 'mon check' category
This is just an embryo so far, but it seems to work ok on a single
node installation. The idea is that we should be able to expand on
this command module by adding checks as commands to it.

The nagios_plugin module helps us create check-commands working as
nagios plugins, so the 'mon check' commands can be run as plugins
as well, complete with perfdata and whatnot.

Note that this currently documents more commands than there are
available. That's intended to be remedied fairly soon.

Also note that this is not at all well-tested, but it will be once
we get this installed in our lab environment.

Signed-off-by: Andreas Ericsson <ae@op5.se>