Add coredump.py, for intelligently examining coredumps
It "knows" about a few binaries that we're particularly interested in
and can help a caller automagically remove cores that are older than
the executables the cores are generated from. This helps quite a lot
when debugging systems that frequently get updated, since old corefiles
need to be discarded under such circumstances.

This is still a work in progress, but at least it's committed somewhere
now.

Signed-off-by: Andreas Ericsson <ae@op5.se>