mon check: Add 'spool' command
Sometimes op5 Monitor systems can suffer a huge buildup of files in
the spool directories. This degrades performance for both perfdata
processing and checkresult processing (the primary audiences for this
check and for which we have some shorthands), but is also a problem for
other programs that uses spool directories.

The default thresholds may need to be tweaked, but this check is useful
as-is right now as well.

Signed-off-by: Andreas Ericsson <ae@op5.se>