mon check: Discount files that disappear from spool directories
Since spool directories are designed to be emptied pretty often it's
not exactly weird that they disappear between the time we parse the
directory contents and the time we stat the files. New files that
have been created between those times are always too new for us to
care about anyway, so we simply discard them.

Signed-off-by: Andreas Ericsson <ae@op5.se>