mon oconf: Assign slave objects based on their name
The code was borrowed from the nagios_service class, and earlier we
left it up to the "service_description" variable to split objects of
a certain typ. Since not all slave objects have that variable even
when they come from the objects.cache file, we use self.name to
assign unique objects to their respective master instead.

Signed-off-by: Andreas Ericsson <ae@op5.se>