Add database support for merlin node connectivity statistics
With this patch, each node can have its own list of how many peers
and pollers are configured and active, as well as each peer's
peer_id and how many checks of each type the node thinks it's
supposed to handled. This will help greatly in determining if
loadbalancing works properly or, if it doesn't, might point to
an obvious error that is timeconsuming and error prone to look
up without this info.

There's still no code support for this feature, but that is up
and coming shortly.

Signed-off-by: Andreas Ericsson <ae@op5.se>