Add (and use) support for sendable parameters in nodeinfo structs
Since we want to track statistics of connected nodes and handled
checks, we need to have support for that in our data structs and
make sure we keep them up to date internally.

The easiest way of making sure they're up to date is to use them for
calculations the same way we do with the previously dominant vars
used for the same purpose but on a single system.

This means the data will now be distributed to all nodes, although
the receivers still has no clue whatsoever what it should do with
the new data.

Signed-off-by: Andreas Ericsson <ae@op5.se>