daemon: Insert node statistics into the database upon arrival
Recent patches to the merlin core added functionality to make nodes
capable of sending their active and configured nodecounts as well
as the peer_id and the number of checks they themselves thought they
were responsible for. This patch takes it one step further and lets
us insert that data into the database for later inspection.

Signed-off-by: Andreas Ericsson <ae@op5.se>