module: Don't reset &self after parsing configuration
Since the patch to store information about how many checks we think
we're supposed to handle (and the update of macros et al to keep
track of it), much of the info therein is calculated when we parse
the configuration. Clobbering that data after the configuration has
been parsed means we'll think we have no other nodes in our network.

Signed-off-by: Andreas Ericsson <ae@op5.se>