daemon: Handle downtime and comment events more or less the same
Since they both use id's internal to Nagios, we need to make sure
to handle them with different heuristics. This has some potentially
nasty sideeffects, such that identical events will be deleted from
the database but possibly not from Nagios itself. That really can't
be helped, and it's highly unlikely that multiple identical events
exist on purpose anyway, so this is probably a sensible fix.

Signed-off-by: Andreas Ericsson <ae@op5.se>