Disambiguate between "broadcast this packet" and CTRL_GENERIC
CTRL_GENERIC is intended for the code field of the merlin protocol
header, while "broadcast this packet" is marked by setting the
selection to 0xffff in the event hook. By adding a new macro named
"DEST_BROADCAST" we can easily distinguish between the two and make
some if() statements slightly easier to read.

Signed-off-by: Andreas Ericsson <ae@op5.se>