module: Send user-triggered notifications from command-receiving node
User-triggered notifications are, in this case, acknowledgements and
"custom" notifications.

This is nifty in order to prevent them from being lost in a crashed
merlin network in case backlog latency is high and the node supposed
to send the command goes offline before its module has time to process
the command. That's a (very) unlikely scenario for the node receiving
the command.

This fixes issue #4498.

Signed-off-by: Andreas Ericsson <ae@op5.se>