module: Revamp downtime and comment handling somewhat
Downtime commands are forwarded as-is, so we must block their comment
events from reaching further than our own daemon. We must, however,
forward DOWNTIME_DELETE and DOWNTIME_STOP events, to make sure the
node responsible for monitoring the host or service we're interested
in doesn't think downtime is still in effect when it actually might
not be.

We also reorder downtime related commands in hook_external_command()
so we can keep track of them a bit easier in case we want to change
how we handle downtime later.

Signed-off-by: Andreas Ericsson <ae@op5.se>