daemon: Refuse to run global config sync command to master nodes
The normal setup is that configuration is done on a master and
then pushed down to pollers from there, but the example config
shipped uses a global push-command that works well for master
to poller and peer to peer but not so great from poller to
master.

This patch prevents peered pollers from engaging in what I'd like
to call "config pingpong", where they both send their newly received
configuration to each other with the global "mon oconf push" command,
restart each other with the same command and then notice that one or
more master nodes have an older configuration so they try to push
again and then again and then again, in absurdum. This patch should
take care of that quite neatly so that the default setup is easy to
manage and the nonstandard ones are possible.

Signed-off-by: Andreas Ericsson <ae@op5.se>