app module merlin_status: Fix assigned_checks() calculation
This was previously broken for any network of peers with pollers
assigned to them and those pollers' assigned checks didn't all end
up on "one side" of the nodes handled by the top-tiered ones from the
alphabetically sorted list of checks.

With this patch, "mon check distribution" works as intended, making it
a very valuable addendum to the suite of autotests performed on the
merlin installation in the lab as well as in production environments.

Signed-off-by: Andreas Ericsson <ae@op5.se>