mon check cores: Force users to pass an option when deleting cores
The previous behaviour of always deleting the corefiles was quite
demanding, since any bugs in it prevented closer inspection. Now we
force the user to pass the --delete flag if they want to delete any
invalid corefiles. Valid corefiles are always kept.

We also add a --debug flag so we can print the reason why certain
corefiles are invalid and thereby catch bugs in the program.

While we're at it with adding new options, we indent the helptext for
them all so it gets a bit easier to read.

Signed-off-by: Andreas Ericsson <ae@op5.se>