db_updater: Never put downtime_delete events in report_data
If the downtime actually started, it will trigger a downtime_end event
as well. By sending both to the database, we get roughly twice as many
downtime end entries as there are downtimes started, which makes it
pretty much impossible to handle nested downtimes for report generation
tools.

This patch behaves as expected when creating one or more overlapping
downtime(s) and leaving it/them until the end, when creating a downtime
and cancelling it halfway through, and when creating a downtime and
cancelling it before it's begun.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>