module: Remove race from net2mod
It's possible, mainly by misconfiguring merlin, to get two nodes in the
same cluster to run the same checks simultaneously. When that happens,
it's possible that one of them will provide the other with a check
result exactly when the other is already saving a result for that check.
If this happens, this patch will close a race condition that can cause
a crash in nagios.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>