Log renames, add new renaming tool
Log nacoma renames in their own database table, and add a tool for
renaming the host/service in database and on file.

The file rewriting is not terribly efficient, however, we really need to
parse all lines fairly extensively to avoid killing our log (if you
rename a host called OK to CRITICAL, bad things shouldn't happen).
I can't really benchmark the actual code performance, though, as my
harddrive is way too slow to keep it busy. That should be a good sign.
Still, it's gonna take a long time to execute, especially on rotating
media.

The database update should be pretty quick, though, as the database has
indexes and stuff so it only has to read/write important things. As the
reports and alert log use the database, you may not even need to do the
whole file rewriting all the time.

This fixes bug #755 for people who run rename fairly regularly.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>