rename: Do nothing by default, options for only current log or archived
Do nothing by default, as this command could be dangerous.

Make it possible to only rename the archive, or only rename the current
log. Also make it possible to not truncate the rename log.

If you're renaming the current log, you need to turn off monitor first.
With these new options, we can now first rename the database and archive
but not truncate, then turn off monitor, then rename the current log and
truncate, then turn monitor back on. Yes, this is becoming quite
complicated, we should move this to mon...

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>