mon log: Add purge command to clean unused RRD files
This fixes bug #3390 for RRD files that became stale in monitor 5.5 or
later, that is, it doesn't work for RRD files that became unused
earlier. The reason for this is that as of 5.5, PNP has a nacoma hook to
keep track of renames and deletions, so it moves unused RRD file to a
trash directory that is easy to find.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>