mon suite: Add 'ecmd' category
This is supremely nifty for finding the various commands one might want
to submit to Nagios, and also for submitting them from the command
line interface. This facilitates testing in combination with the
immensely nifty 'mon node ctrl' command, since complicated echo strings
need not be remembered.

No checks are done to see if the hosts and services we're trying to
submit commands for actually exist. That's left as an exercise to the
people writing scripts around this tool.

"ecmd" stands for "easy/external commands", in case anyone's curious.

Signed-off-by: Andreas Ericsson <ae@op5.se>