Introduce 'mon test dist'
This nifty little helper launches a bunch of nagios+merlin instances
on the local host where they're set to communicate with each other
and a single process can easily submit commands to and check the
databases of all the instances supremely simply.

This will be the foundation of all tests we need to write to fix
communications-related issues with downtime, acknowledgements, comments
and whatnot.

Signed-off-by: Andreas Ericsson <ae@op5.se>